refactor(types): move EventEmitter trait + NoopEmitter to origin-types
