+++ crates/origin-types/Cargo.toml
@@ -12,5 +12,6 @@ keywords = ["memory", "mcp", "origin", "agent"]
 categories = ["data-structures", "api-bindings"]
 
 [dependencies]
+anyhow = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
